Move the data components up a level
