+++ docs/src/pages/tables/index.jsx
@@ -1,12 +1,18 @@
 import { UsaceBox } from "@usace/groundwork";
 import {TSTable} from "@usace-watermanagement/groundwork-water";
-import DocsPage from "../_docs-wrapper";
+import { Code } from "../components/code";
+import Divider from "../components/divider";
+import DocsPage from "../docs/_docs-wrapper";
 
 function Tables() {
   return (
-    <DocsPage>
+    <DocsPage
+    nextUrl="/docs/maps"
+    prevUrl="/docs/plots"
+    middleText="Tables"
+    prevText="Return to Plots"
+    nextText="Go to Maps Page">
     <div className="mt-6">
-      <UsaceBox title="Welcome">
         <TSTable
           queryParams={{
             office: "SWT",
@@ -14,7 +20,6 @@ function Tables() {
           }}
           precision={2}
         />
-      </UsaceBox>
     </div>
     </DocsPage>
   );
